# api/server/docs/swagger.yaml
basePath: /
consumes:
- application/json
...
info:
  contact:
    email: support@e154.ru
    name: Alex Filippov
    url: https://e154.github.io/smart-home-gate/
  description: This documentation describes APIs found under https://github.com/e154/smart-home-gate
  license:
    name: MIT
    url: https://raw.githubusercontent.com/e154/smart-home-gate/master/LICENSE
  title: Smart home gate
  version: 1.0.0
paths:
...
  /check/mobile_access_token:
    get:
      consumes:
      - text/plain
      operationId: check
      produces:
      - text/plain
      responses:
        "200":
          description: Success response
        "400":
          description: Bad request
        "404":
          description: Not found
      security:
      - ServerAuthorization: []
      summary: mobile access token check page
      tags:
      - check
...
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ServerAuthorization:
    in: header
    name: ServerAuthorization
    type: apiKey
swagger: "2.0"